Use SDL3 Wayland on Linux; Steam Deck fixes
Switch Linux packages and docs to default to the SDL3 backend and enable native Wayland via SDL3/Wayland/EGL. Update Steam Deck launcher to only force XWayland when no SDL_VIDEO_DRIVER/SDL_VIDEODRIVER is set. Add desktop launcher installer and several VMware helper scripts. Expand joystick mappings in default.cfg and add new HUD GUI assets. Update numerous platform, SDL3, meson and build scripts/source files to support the SDL3 migration and platform improvements.

+++ src/bse/BSE_Particle.cpp
@@ -1831,10 +1831,10 @@ bool rvElectricityParticle::Render(const rvBSE* effect, rvParticleTemplate* pt, 
 
 	if (mLastJitter + mJitterRate <= time) {
 		mLastJitter = time;
-		mSeed = static_cast<unsigned long>(rvRandom::Init());
+		mSeed = rvRandom::Init();
 	}
 
-	const unsigned long previousSeed = rvRandom::GetSeed();
+	const unsigned int previousSeed = rvRandom::GetSeed();
 	rvRandom::Init(mSeed);
 
 	const int boltCount = Max(1, (mNumBolts > 0) ? mNumBolts : GetBoltCount(mainLength));
@@ -1916,7 +1916,7 @@ void rvElectricityParticle::SetupElectricity(rvParticleTemplate* pt) {
 	const rvElectricityInfo* info = pt->mElecInfo;
 	mNumBolts = 0;
 	mNumForks = info->mNumForks;
-	mSeed = static_cast<unsigned long>(rvRandom::Init());
+	mSeed = rvRandom::Init();
 	mForkSizeMins = info->mForkSizeMins;
 	mForkSizeMaxs = info->mForkSizeMaxs;
 	mJitterSize = info->mJitterSize;
